+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Clear possible old user feedback event before saving image
+- Return meaningful PostcardScannerState
 
 ## [1.4.0] - 2023-09-25
 ### Added